move merkle commiter
