+++ ledger/store/accountsV2.go
@@ -458,3 +458,9 @@ func (w *accountsV2Writer) UpdateAccountsHashRound(ctx context.Context, hashRoun
 	}
 	return
 }
+
+// ResetAccountHashes resets the account hashes generated by the merkle commiter.
+func (w *accountsV2Writer) ResetAccountHashes(ctx context.Context) (err error) {
+	_, err = w.e.ExecContext(ctx, `DELETE FROM accounthashes`)
+	return
+}
