+++ ledger/store/merkle_commiter.go
@@ -1,0 +1,75 @@
+// Copyright (C) 2019-2022 Algorand, Inc.
+// This file is part of go-algorand
+//
+// go-algorand is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// go-algorand is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with go-algorand.  If not, see <https://www.gnu.org/licenses/>.
+
+package store
+
+import "database/sql"
+
+// MerkleCommitter allows storing and loading merkletrie pages from a sqlite database.
+//
+//msgp:ignore MerkleCommitter
+type MerkleCommitter struct {
+	tx         *sql.Tx
+	deleteStmt *sql.Stmt
+	insertStmt *sql.Stmt
+	selectStmt *sql.Stmt
+}
+
+// MakeMerkleCommitter creates a MerkleCommitter object that implements the merkletrie.Committer interface allowing storing and loading
+// merkletrie pages from a sqlite database.
+func MakeMerkleCommitter(tx *sql.Tx, staging bool) (mc *MerkleCommitter, err error) {
+	mc = &MerkleCommitter{tx: tx}
+	accountHashesTable := "accounthashes"
+	if staging {
+		accountHashesTable = "catchpointaccounthashes"
+	}
+	mc.deleteStmt, err = tx.Prepare("DELETE FROM " + accountHashesTable + " WHERE id=?")
+	if err != nil {
+		return nil, err
+	}
+	mc.insertStmt, err = tx.Prepare("INSERT OR REPLACE INTO " + accountHashesTable + "(id, data) VALUES(?, ?)")
+	if err != nil {
+		return nil, err
+	}
+	mc.selectStmt, err = tx.Prepare("SELECT data FROM " + accountHashesTable + " WHERE id = ?")
+	if err != nil {
+		return nil, err
+	}
+	return mc, nil
+}
+
+// StorePage is the merkletrie.Committer interface implementation, stores a single page in a sqlite database table.
+func (mc *MerkleCommitter) StorePage(page uint64, content []byte) error {
+	if len(content) == 0 {
+		_, err := mc.deleteStmt.Exec(page)
+		return err
+	}
+	_, err := mc.insertStmt.Exec(page, content)
+	return err
+}
+
+// LoadPage is the merkletrie.Committer interface implementation, load a single page from a sqlite database table.
+func (mc *MerkleCommitter) LoadPage(page uint64) (content []byte, err error) {
+	err = mc.selectStmt.QueryRow(page).Scan(&content)
+	if err == sql.ErrNoRows {
+		content = nil
+		err = nil
+		return
+	} else if err != nil {
+		return nil, err
+	}
+	return content, nil
+}
